Changed comma_counter to size_t

+++ compiler_src/comma_input.c
@@ -6,7 +6,7 @@
 static char *comma_inputs;
 static size_t comma_index = 0;
 
-void get_comma_inputs(int comma_counter) {
+void get_comma_inputs(size_t comma_counter) {
     if (comma_counter <= 0) {
         comma_inputs = NULL;
         return;
@@ -18,12 +18,12 @@ void get_comma_inputs(int comma_counter) {
         exit(EXIT_FAILURE);
     }
 
-    printf("Comma inputs (Provide %d character%s for ',' commands):\n", comma_counter, (comma_counter == 1 ? "" : "s"));
+    printf("Comma inputs (Provide %ld character%s for ',' commands):\n", comma_counter, (comma_counter == 1 ? "" : "s"));
 
     size_t comma_index = 0;
     char line[1024];
 
-    while (comma_index < (size_t) comma_counter) {
+    while (comma_index < comma_counter) {
         if (fgets(line, sizeof(line), stdin) == NULL) {
             fprintf(stderr, "Error reading input.\n");
             exit(EXIT_FAILURE);
@@ -35,9 +35,9 @@ void get_comma_inputs(int comma_counter) {
             }
         }
 
-        if (comma_index < (size_t) comma_counter) {
-            int commas_left = comma_counter - (int) comma_index;
-            printf("\nComma inputs (Provide %d more character%s for ',' commands):\n", commas_left, (commas_left == 1 ? "" : "s"));
+        if (comma_index < comma_counter) {
+            size_t commas_left = comma_counter - comma_index;
+            printf("\nComma inputs (Provide %ld more character%s for ',' commands):\n", commas_left, (commas_left == 1 ? "" : "s"));
         }
     }
     printf("\n");
@@ -45,7 +45,7 @@ void get_comma_inputs(int comma_counter) {
     printf("Output:\n");
 }
 
-char next_comma_input(int comma_counter) {
+char next_comma_input(size_t comma_counter) {
     if (comma_index >= comma_counter) {
         fprintf(stderr, "Runtime error: Comma input not found.\n");
         exit(EXIT_FAILURE);
@@ -54,7 +54,7 @@ char next_comma_input(int comma_counter) {
     return comma_inputs[comma_index++];
 }
 
-void free_comma_inputs(int comma_counter) {
+void free_comma_inputs(size_t comma_counter) {
     if (comma_counter > 0) {
         free(comma_inputs);
     }
